Naive draft of adding rolls
Now you can roll, but the application does not help you with
useful feedback yet.

In order to be of any reasonable help, we must also judge if the outcome
of the characteristic roll was successful - or not.

This shall be our next step. We must come back to this later once proven
that this draft feels nice when using it.
Apart from newly introduced TODOs and FIXMEs, the matter of tests must
be addressed. We will probably delete this implementation later to come
up with a nicer TDD design.